Move mouseInteractionObserver
