+++ packages/rum/src/domain/record/observers/mouseInteractionObserver.spec.ts
@@ -1,0 +1,145 @@
+import { DefaultPrivacyLevel, isIE, noop } from '@datadog/browser-core'
+import { createNewEvent } from '@datadog/browser-core/test/specHelper'
+import type { RumConfiguration } from '@datadog/browser-rum-core'
+import { IncrementalSource, MouseInteractionType, RecordType } from '../../../types'
+import { serializeDocument, SerializationContextStatus } from '../serialize'
+import { createElementsScrollPositions } from '../elementsScrollPositions'
+import { getRecordIdForEvent } from '../utils'
+import type { ShadowRootsController } from '../shadowRootsController'
+import { NodePrivacyLevel } from '../../../constants'
+import type { MouseInteractionCallBack } from './mouseInteractionObserver'
+import { initMouseInteractionObserver } from './mouseInteractionObserver'
+
+const DEFAULT_SHADOW_ROOT_CONTROLLER: ShadowRootsController = {
+  flush: noop,
+  stop: noop,
+  addShadowRoot: noop,
+  removeShadowRoot: noop,
+}
+
+const DEFAULT_CONFIGURATION = { defaultPrivacyLevel: NodePrivacyLevel.ALLOW } as RumConfiguration
+
+describe('initMouseInteractionObserver', () => {
+  let mouseInteractionCallbackSpy: jasmine.Spy<MouseInteractionCallBack>
+  let stopObserver: () => void
+  let sandbox: HTMLDivElement
+  let a: HTMLAnchorElement
+
+  beforeEach(() => {
+    if (isIE()) {
+      pending('IE not supported')
+    }
+
+    sandbox = document.createElement('div')
+    a = document.createElement('a')
+    a.setAttribute('tabindex', '0') // make the element focusable
+    sandbox.appendChild(a)
+    document.body.appendChild(sandbox)
+    a.focus()
+
+    serializeDocument(document, DEFAULT_CONFIGURATION, {
+      shadowRootsController: DEFAULT_SHADOW_ROOT_CONTROLLER,
+      status: SerializationContextStatus.INITIAL_FULL_SNAPSHOT,
+      elementsScrollPositions: createElementsScrollPositions(),
+    })
+
+    mouseInteractionCallbackSpy = jasmine.createSpy()
+    stopObserver = initMouseInteractionObserver(mouseInteractionCallbackSpy, DefaultPrivacyLevel.ALLOW)
+  })
+
+  afterEach(() => {
+    sandbox.remove()
+    stopObserver()
+  })
+
+  it('should generate click record', () => {
+    a.click()
+
+    expect(mouseInteractionCallbackSpy).toHaveBeenCalledWith({
+      id: jasmine.any(Number),
+      type: RecordType.IncrementalSnapshot,
+      timestamp: jasmine.any(Number),
+      data: {
+        source: IncrementalSource.MouseInteraction,
+        type: MouseInteractionType.Click,
+        id: jasmine.any(Number),
+        x: jasmine.any(Number),
+        y: jasmine.any(Number),
+      },
+    })
+  })
+
+  it('should generate mouseup record on pointerup DOM event', () => {
+    const pointerupEvent = createNewEvent('pointerup', { clientX: 1, clientY: 2 })
+    a.dispatchEvent(pointerupEvent)
+
+    expect(mouseInteractionCallbackSpy).toHaveBeenCalledWith({
+      id: getRecordIdForEvent(pointerupEvent),
+      type: RecordType.IncrementalSnapshot,
+      timestamp: jasmine.any(Number),
+      data: {
+        source: IncrementalSource.MouseInteraction,
+        type: MouseInteractionType.MouseUp,
+        id: jasmine.any(Number),
+        x: jasmine.any(Number),
+        y: jasmine.any(Number),
+      },
+    })
+  })
+
+  it('should not generate click record if x/y are missing', () => {
+    const clickEvent = createNewEvent('click')
+    a.dispatchEvent(clickEvent)
+
+    expect(mouseInteractionCallbackSpy).not.toHaveBeenCalled()
+  })
+
+  it('should generate blur record', () => {
+    a.blur()
+
+    expect(mouseInteractionCallbackSpy).toHaveBeenCalledWith({
+      id: jasmine.any(Number),
+      type: RecordType.IncrementalSnapshot,
+      timestamp: jasmine.any(Number),
+      data: {
+        source: IncrementalSource.MouseInteraction,
+        type: MouseInteractionType.Blur,
+        id: jasmine.any(Number),
+      },
+    })
+  })
+
+  // related to safari issue, see RUMF-1450
+  describe('forced layout issue', () => {
+    let coordinatesComputed: boolean
+
+    beforeEach(() => {
+      if (!window.visualViewport) {
+        pending('no visualViewport')
+      }
+
+      coordinatesComputed = false
+      Object.defineProperty(window.visualViewport, 'offsetTop', {
+        get() {
+          coordinatesComputed = true
+          return 0
+        },
+        configurable: true,
+      })
+    })
+
+    afterEach(() => {
+      delete (window.visualViewport as any).offsetTop
+    })
+
+    it('should compute x/y coordinates for click record', () => {
+      a.click()
+      expect(coordinatesComputed).toBeTrue()
+    })
+
+    it('should not compute x/y coordinates for blur record', () => {
+      a.blur()
+      expect(coordinatesComputed).toBeFalse()
+    })
+  })
+})
