+++ packages/rum/src/domain/record/observers/observers.spec.ts
@@ -4,18 +4,13 @@ import { ActionType, LifeCycle, LifeCycleEventType, RumEventType, FrustrationTyp
 import type { RawRumEventCollectedData } from 'packages/rum-core/src/domain/lifeCycle'
 import { createNewEvent, isFirefox } from '@datadog/browser-core/test/specHelper'
 import { NodePrivacyLevel, PRIVACY_ATTR_NAME, PRIVACY_ATTR_VALUE_MASK_USER_INPUT } from '../../../constants'
-import { IncrementalSource, MouseInteractionType, RecordType } from '../../../types'
+import { RecordType } from '../../../types'
 import { serializeDocument, SerializationContextStatus } from '../serialize'
 import { createElementsScrollPositions } from '../elementsScrollPositions'
 import type { ShadowRootsController } from '../shadowRootsController'
-import {
-  getRecordIdForEvent,
-  initStyleSheetObserver,
-  initFrustrationObserver,
-  initInputObserver,
-  initMouseInteractionObserver,
-} from './observers'
-import type { FrustrationCallback, InputCallback, MouseInteractionCallBack, StyleSheetCallback } from './observers'
+import { getRecordIdForEvent } from '../utils'
+import { initStyleSheetObserver, initFrustrationObserver, initInputObserver } from './observers'
+import type { FrustrationCallback, InputCallback, StyleSheetCallback } from './observers'
 
 const DEFAULT_SHADOW_ROOT_CONTROLLER: ShadowRootsController = {
   flush: noop,
@@ -332,131 +327,6 @@ describe('initStyleSheetObserver', () => {
 
         expect(styleSheetCallbackSpy).not.toHaveBeenCalled()
       })
-    })
-  })
-})
-
-describe('initMouseInteractionObserver', () => {
-  let mouseInteractionCallbackSpy: jasmine.Spy<MouseInteractionCallBack>
-  let stopObserver: () => void
-  let sandbox: HTMLDivElement
-  let a: HTMLAnchorElement
-
-  beforeEach(() => {
-    if (isIE()) {
-      pending('IE not supported')
-    }
-
-    sandbox = document.createElement('div')
-    a = document.createElement('a')
-    a.setAttribute('tabindex', '0') // make the element focusable
-    sandbox.appendChild(a)
-    document.body.appendChild(sandbox)
-    a.focus()
-
-    serializeDocument(document, DEFAULT_CONFIGURATION, {
-      shadowRootsController: DEFAULT_SHADOW_ROOT_CONTROLLER,
-      status: SerializationContextStatus.INITIAL_FULL_SNAPSHOT,
-      elementsScrollPositions: createElementsScrollPositions(),
-    })
-
-    mouseInteractionCallbackSpy = jasmine.createSpy()
-    stopObserver = initMouseInteractionObserver(mouseInteractionCallbackSpy, DefaultPrivacyLevel.ALLOW)
-  })
-
-  afterEach(() => {
-    sandbox.remove()
-    stopObserver()
-  })
-
-  it('should generate click record', () => {
-    a.click()
-
-    expect(mouseInteractionCallbackSpy).toHaveBeenCalledWith({
-      id: jasmine.any(Number),
-      type: RecordType.IncrementalSnapshot,
-      timestamp: jasmine.any(Number),
-      data: {
-        source: IncrementalSource.MouseInteraction,
-        type: MouseInteractionType.Click,
-        id: jasmine.any(Number),
-        x: jasmine.any(Number),
-        y: jasmine.any(Number),
-      },
-    })
-  })
-
-  it('should generate mouseup record on pointerup DOM event', () => {
-    const pointerupEvent = createNewEvent('pointerup', { clientX: 1, clientY: 2 })
-    a.dispatchEvent(pointerupEvent)
-
-    expect(mouseInteractionCallbackSpy).toHaveBeenCalledWith({
-      id: getRecordIdForEvent(pointerupEvent),
-      type: RecordType.IncrementalSnapshot,
-      timestamp: jasmine.any(Number),
-      data: {
-        source: IncrementalSource.MouseInteraction,
-        type: MouseInteractionType.MouseUp,
-        id: jasmine.any(Number),
-        x: jasmine.any(Number),
-        y: jasmine.any(Number),
-      },
-    })
-  })
-
-  it('should not generate click record if x/y are missing', () => {
-    const clickEvent = createNewEvent('click')
-    a.dispatchEvent(clickEvent)
-
-    expect(mouseInteractionCallbackSpy).not.toHaveBeenCalled()
-  })
-
-  it('should generate blur record', () => {
-    a.blur()
-
-    expect(mouseInteractionCallbackSpy).toHaveBeenCalledWith({
-      id: jasmine.any(Number),
-      type: RecordType.IncrementalSnapshot,
-      timestamp: jasmine.any(Number),
-      data: {
-        source: IncrementalSource.MouseInteraction,
-        type: MouseInteractionType.Blur,
-        id: jasmine.any(Number),
-      },
-    })
-  })
-
-  // related to safari issue, see RUMF-1450
-  describe('forced layout issue', () => {
-    let coordinatesComputed: boolean
-
-    beforeEach(() => {
-      if (!window.visualViewport) {
-        pending('no visualViewport')
-      }
-
-      coordinatesComputed = false
-      Object.defineProperty(window.visualViewport, 'offsetTop', {
-        get() {
-          coordinatesComputed = true
-          return 0
-        },
-        configurable: true,
-      })
-    })
-
-    afterEach(() => {
-      delete (window.visualViewport as any).offsetTop
-    })
-
-    it('should compute x/y coordinates for click record', () => {
-      a.click()
-      expect(coordinatesComputed).toBeTrue()
-    })
-
-    it('should not compute x/y coordinates for blur record', () => {
-      a.blur()
-      expect(coordinatesComputed).toBeFalse()
     })
   })
 })
